frontend/requirements.txt
Flask==2.2.3
python-dotenv==0.15.0
Flask-Login==0.6.2